Feature Final:
A state part for books with an array of books (initial state: empty array), reducer that adds a book, and reducer that removes an added book. Added a status slice for categories with an array of categories (initial state: empty array) and a reducer that checks for status and always returns "Under Construction" (initial state should check for that string).
Structure of the folders followed:
/src
     |---/components
     |---/redux
         |--- /books
                 | booksSlice.js
         |--- /categories
                 | categoriesSlice.js
         | store.js
     | index.js